Enemy finished, sound added
Changed a lot of stuff about how the enemy works, made it so that its error is increased by the speed of the ball. A smart thing to do would be to again increase the error by the distance the enemy needs to travel, I might do this.

Sound has started and it will be pretty simple to finish.