Controller support finished, game finished
Pong is in a very solid place now.
I'm considering adding a higher margin of error the farther that the enemy has to go, but that will be a small update in the future if so, I'm considering the game done.

The game has:
- Enemy AI that uses raycasts to find where to go
- Single player against AI
- Local multiplayer
- Full controller support
- Sound that I will credit in the README

Overall, a pretty simple game that taught me a lot.